contactReducer.js: Added UPDATE_CONTACT switch case
The map method returns a new array from running a function on each item. In this case if the item ID is the same as the payload ID we pass in, we do not add the contact currently in state and instead add the payload, because that is the updated contact.